requirements.txt
aiofiles==22.1.0
altair==4.2.0
anyio==3.6.2
attrs==22.2.0
autopep8==2.0.1
beautifulsoup4==4.11.1
blinker==1.5
cachetools==5.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
ci-info==0.3.0
click==8.1.3
click-plugins==1.1.1
cohere==3.1.5
commonmark==0.9.1
configobj==5.0.6
configparser==5.3.0
decorator==5.1.1
docopt==0.6.2
entrypoints==0.4
etelemetry==0.3.0
filelock==3.9.0
fritz==2.3.5
frontend==0.0.3
future==0.18.3
gdown==4.6.0
gitdb==4.0.10
GitPython==3.1.30
h11==0.14.0
httplib2==0.21.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
looseversion==1.0.3
lxml==4.9.2
MarkupSafe==2.1.1
networkx==3.0
nibabel==5.0.0
nipype==1.8.5
numpy==1.24.1
openstep-parser==1.5.4
packaging==23.0
pandas==1.5.2
pathlib==1.0.1
pbxproj==3.5.0
pdf2image==1.16.2
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
protobuf==3.20.3
prov==2.0.0
pyarrow==10.0.1
pybind11==2.10.3
pycodestyle==2.10.0
pydeck==0.8.0
pydot==1.4.2
Pygments==2.14.0
Pympler==1.0.1
PyMuPDF==1.21.1
pyparsing==3.0.9
pyperclip==1.8.2
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pyxnat==1.5
rdflib==6.2.0
requests==2.28.2
rich==13.1.0
scipy==1.10.0
semver==2.13.0
simplejson==3.18.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
starlette==0.23.1
streamlit==1.17.0
termcolor==2.2.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traits==6.3.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
uvicorn==0.20.0
validators==0.20.0
watchdog==2.2.1
zipp==3.11.0
